refactor logic out of data_handler
