+++ backend/etl/data_handler.py
@@ -1,7 +1,5 @@
 from abc import ABC, abstractmethod
 import requests
-from requests.adapters import HTTPAdapter
-from urllib3.util.retry import Retry
 from sqlalchemy.dialects.postgresql import insert as pg_insert
 from backend.database.session import get_db
 from backend.api.models.base import ModelType
@@ -11,6 +9,8 @@ from typing import Type, Generator, Optional
 import time
 import logging
 from urllib.parse import unquote
+from .session_manager import SessionManager
+from .request_handler import RequestHandler
 
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
@@ -44,6 +44,9 @@ class DataHandler(ABC):
         self.page_size = page_size
         self.params = params or {}
         self.logger = logger or logging.getLogger(f"{self.__class__.__name__}")
+        self.session = session or SessionManager.create_session(self.logger)
+        self.request_handler = RequestHandler(self.session, self.logger)
+
         self.logger.info(
             f"Initialized handler for {table.__name__} "
             f"with URL: {url}, "
@@ -51,175 +54,6 @@ class DataHandler(ABC):
             f"params: {params}, "
             f"session: {session}"
         )
-        if session is None:
-            session = self._create_session()
-        self.session = session
-
-    def _create_session(self) -> requests.Session:
-        """Create a configured requests session with retry logic."""
-
-        class LoggingRetry(Retry):
-            def __init__(self, *args, **kwargs):
-                super().__init__(*args, **kwargs)
-                self.logger = logging.getLogger("RetryLogger")
-                self.logger.setLevel(logging.INFO)
-                self._max_allowed_retries = kwargs.get("total", 0)
-
-            def increment(
-                self, method=None, url=None, response=None, error=None, *args, **kwargs
-            ):
-                """Called when a retry is needed"""
-                # Check if we've exhausted our retries
-                if self.total <= 0:
-                    self.logger.error(
-                        f"Max retries ({self._max_allowed_retries}) exceeded. Giving up."
-                    )
-                    return super().increment(
-                        method=method,
-                        url=url,
-                        response=response,
-                        error=error,
-                        *args,
-                        **kwargs,
-                    )
-
-                current_attempt = self._max_allowed_retries - self.total + 1
-
-                if response and hasattr(response, "status_code"):
-                    status = response.status_code
-                else:
-                    status = "unknown"
-
-                backoff = self.get_backoff_time()
-
-                self.logger.warning(
-                    f"""
-                            === Retry Attempt {current_attempt} of {self._max_allowed_retries} ===
-                            URL: {unquote(url)}
-                            Method: {method}
-                            Status: {status}
-                            Error: {str(error) if error else "no error"}
-                            Backoff: {backoff} seconds
-                    """
-                )
-
-                # Sleep for backoff duration
-                if backoff:
-                    time.sleep(backoff)
-
-                # Call parent's increment to handle the retry logic
-                return super().increment(
-                    method=method,
-                    url=url,
-                    response=response,
-                    error=error,
-                    *args,
-                    **kwargs,
-                )
-
-            def get_backoff_time(self):
-                """
-                Calculate the backoff time for the current retry attempt.
-                Using *args to maintain compatibility with parent class calls.
-                """
-                current_attempt = self._max_allowed_retries - self.total
-                if current_attempt <= 0:
-                    return 0
-                # Calculate exponential backoff: 0, 2, 4, 8, 16 seconds
-                return self.backoff_factor * (2 ** (current_attempt))
-
-            def new(self, **kw):
-                """Preserve initial total when copying the retry object"""
-                obj = super().new(**kw)
-                obj._max_allowed_retries = self._max_allowed_retries
-                return obj
-
-        retry_strategy = LoggingRetry(
-            total=5,
-            backoff_factor=1,
-            status_forcelist=[
-                404,  # Not Found
-                429,  # Too Many Requests
-                500,  # Internal Server Error
-                502,  # Bad Gateway
-                503,  # Service Unavailable
-                504,  # Gateway Timeout
-            ],
-            allowed_methods=["GET"],
-            raise_on_status=True,
-            respect_retry_after_header=True,
-        )
-
-        # Create session with retry strategy
-        session = requests.Session()
-        adapter = HTTPAdapter(max_retries=retry_strategy)
-        session.mount("https://", adapter)
-        session.mount("http://", adapter)
-
-        self.logger.info(
-            f"Created new session with retry configuration:\n"
-            f"Max retries: {retry_strategy.total}\n"
-            f"Backoff factor: {retry_strategy.backoff_factor}\n"
-            f"Status codes: {retry_strategy.status_forcelist}\n"
-            f"Allowed methods: {retry_strategy.allowed_methods}\n"
-            f"Respect retry after header: {retry_strategy.respect_retry_after_header}"
-        )
-
-        return session
-
-    def _make_request(self, url: str, params: dict, timeout: int = 60) -> dict:
-        """
-        Make HTTP request with error handling.
-        Args:
-            url: Request URL
-            params: Query parameters
-            timeout: Request timeout in seconds
-        Returns:
-            Response data as dictionary containing list of features under 'features' key.
-        """
-        start_time = time.time()
-        try:
-            self.logger.info(f"Making request to {url} with params {params}")
-
-            response = self.session.get(url, params=params, timeout=timeout)
-            self.logger.info(f"Got response with status: {response.status_code}")
-
-            response.raise_for_status()
-            try:
-                data = response.json()
-            except Exception as e:
-                self.logger.error(f"Failed to parse JSON: {str(e)}")
-                raise
-
-            self.logger.info(
-                f"Request completed successfully:\n"
-                f"URL: {url}\n"
-                f"Params: {params}\n"
-                f"Time: {time.time() - start_time:.2f}s"
-            )
-
-            return data
-
-        except requests.HTTPError as e:
-            self.logger.error(
-                f"HTTP Error occurred:\n"
-                f"Status Code: {e.response.status_code}\n"
-                f"URL: {url}\n"
-                f"Params: {params}\n"
-                f"Time: {time.time() - start_time:.2f}s",
-                exc_info=True,
-            )
-            raise
-        except Exception as e:
-            self.logger.error(
-                f"Request failed:\n"
-                f"Error: {str(e)}\n"
-                f"URL: {url}\n"
-                f"Params: {params}\n"
-                f"Time: {time.time() - start_time:.2f}s",
-                exc_info=True,
-            )
-            raise
 
     def _yield_data(self) -> Generator:
         """
@@ -235,9 +69,10 @@ class DataHandler(ABC):
             paginated_params.update({"$offset": offset, "$limit": self.page_size})
 
             start_time = time.time()
-            data = self._make_request(self.url, paginated_params)
+            data = self.request_handler.make_request(self.url, paginated_params)
             features = data.get("features", [])
             request_time = time.time() - start_time
+
             if not features:
                 self.logger.info(
                     f"{self.table.__name__}: Completed pagination. "
